Bump the version on hackage.
Since it was last tagged 1.2.0.1, this package has not had any breaking changes.
It has had one important fix (#60), but this has not been pushed to hackage.

As a consequence hi, does not build on hackage (and also on NixOS) and `cabal
install` fails.

This change adds a new minor version which includes the fix in #60.